requirements.txt
tabula-py==2.9.0
pandas==2.2.0
PyPDF2==3.0.1
openpyxl==3.1.2
pdfplumber==0.10.3
numpy==1.26.3
Flask==3.0.0
Werkzeug==3.0.1
gunicorn==21.2.0
pytest==7.4.3
pytest-flask==1.3.0